refactor: renames and makes the component dynamic
